Fix import issues and restructure files
Fix import issues, merge duplicate files, and restructure the repository for better organization.

* **Import Issues:**
  - Fix import path for `AuthRoute` in `src/App.jsx`.
  - Fix import path for `Dashboard` in `src/App.jsx`.

* **Merge Duplicate Files:**
  - Merge content from `src/components/AuthRoute.tsx` into `src/components/auth/AuthRoute.tsx`.
  - Delete `src/components/AuthRoute.tsx`.
  - Delete `public/JSX/AuthRoute.jsx`.

* **Restructure Repository:**
  - Rename and merge `src/components/layout/styles/theme.js` into `src/theme/theme.js`.
  - Update `README.md` to reflect the new structure and changes made.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Geaux-Specialist-L-L-C/GA-MVP?shareId=XXXX-XXXX-XXXX-XXXX).
