+++ src/components/auth/AuthRoute.tsx
@@ -1,22 +1,17 @@
-import React from 'react';
-import { Navigate, Outlet } from 'react-router-dom';
+import { Navigate, Outlet, useLocation } from 'react-router-dom';
 import { useAuth } from '../../contexts/AuthContext';
 import LoadingSpinner from '../common/LoadingSpinner';
 
-interface AuthRouteProps {}
-
-const AuthRoute: React.FC<AuthRouteProps> = () => {
+const AuthRoute = () => {
   const { currentUser, loading } = useAuth();
-
-  console.log('AuthRoute - currentUser:', currentUser);
-  console.log('AuthRoute - loading:', loading);
+  const location = useLocation();
 
   if (loading) {
     return <LoadingSpinner />;
   }
 
   if (!currentUser) {
-    return <Navigate to="/login" replace />;
+    return <Navigate to="/login" state={{ from: location }} replace />;
   }
 
   if (location.pathname === '/dashboard' && !currentUser.profileComplete) {
