--- src/components/layout/styles/theme.js
@@ -1,17 +1,0 @@
-const theme = {
-  colors: {
-    primary: '#646cff',
-    secondary: '#535bf2',
-    text: '#213547',
-    background: '#ffffff',
-    error: '#ff3333',
-    success: '#4caf50'
-  },
-  breakpoints: {
-    mobile: '480px',
-    tablet: '768px',
-    desktop: '1024px'
-  }
-};
-
-export default theme;
